+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 
+### [1.3.25] - Unreleased
+- Add `WidgetExt::from_dyn_widget()`.
+- Add `utils::is<WidgetExt>(w) and is_ptr_of()`.
+- Add `WidgetExt::image_ref() and deimage_ref()`.
+- Mark `WidgetExt::as_widget_ptr() and ImageExt::as_image_ptr()` as safe.
+- Pull FLTK fixes.
+
 ### [1.3.24] - 2022-12-15
 - Add image::IcoImage.
 - Use Fl_Widget::bind_image and bind_deimage internally.